Refactor recommended-test
This aims to check that the bad fixtures are failing as we intend. Some
fixtures are failing, but not related to the rule we want to test.

As an example, the following snippet:

```js
// require-spread/bad/example.js
const wrap = (f, g) => {
  return (...args) => g.apply(g, [f].concat(args));
}
```

The intention is for this snippet to fail because it should be
doing `g(f, ...args)`, but it is really failing because of the lack
of a `;`.